generating shapes instead of letters
