+++ Makefile
@@ -28,6 +28,10 @@ generate_mnist:
 generate_curve_lines: generate_backgrounds
 	$(PYTHON) generate_curve_lines.py
 
+# Generate shape dataset
+generate_shape_dataset: generate_backgrounds
+	$(PYTHON) generate_shape_dataset.py
+
 # Compare video frames with MNIST samples
 compare_samples:
 	$(PYTHON) compare_samples.py
@@ -42,11 +46,8 @@ copy_config:
 		shutil.copy('config.yaml', os.path.join(base_dir, 'config.yaml'))"
 
 # Generate complete dataset
-generate_dataset: venv copy_config generate_backgrounds generate_curve_lines generate_videos generate_masked_videos compare_samples create_video_index
+generate_dataset: venv copy_config generate_backgrounds generate_curve_lines generate_shape_dataset generate_videos generate_masked_videos compare_samples create_video_index
 
 # Create video index
 create_video_index:
 	$(PYTHON) create_video_index.py
-
-
-.PHONY: venv generate_backgrounds generate_videos generate_masked_videos generate_mnist generate_curve_lines compare_samples generate_dataset create_video_index copy_config
